scripts/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.1.3
numpy>=1.26.0
pyarrow>=15
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
scipy>=1.12.0
statsmodels>=0.14
jupyter>=1.0.0
jupytext>=1.16
ipython>=8.17.2
pytz>=2023.3
stravalib>=0.10.2
garminconnect>=0.3.3
